requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
django-extensions==3.2.1
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
Pillow==9.4.0
promise==2.3
pycodestyle==2.10.0
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
